Get Batches
Implement get_batches to create batches of input and targets using
int_text. The batches should be a Numpy array with the shape (number of
batches, 2, batch size, sequence length). Each batch contains two
elements:
The first element is a single batch of input with the shape [batch
size, sequence length]
The second element is a single batch of targets with the shape [batch
size, sequence length]